fix(model): unified_forward Q layout + f16 byte-offset bugs
Two bugs that produced garbage output on the paged-CUDA smoke test
(prompt of 16 tokens → token salad):

1. **Q layout mismatch.**
   `split_qkv_norm_rope_into_paged_cache_f16` was writing Q in
   head-major `[q_heads, tokens, hd]` (matching the variable name
   `q_head_major` in the caller). `paged_varlen_attention.cu` reads Q
   as token-major `[total_q, num_heads, hd]`. For tokens=1 the two
   layouts collapse, so the Phase 4b batched decode that already
   shipped on Metal was unaffected; the new chunked-prefill path with
   tokens>1 read interleaved garbage.

   Switched the kernel's Q write to token-major. Phase 4b stays
   correct because it always passes tokens=1.

2. **`size_of::<f32>()` for f16 buffers.**
   Both call sites (Phase 4b batched + new unified_forward) computed
   per-item byte offsets as `q_dim * size_of::<f32>()` even though
   `B::Buffer` is `CudaSlice<f16>` / Metal half (2 bytes/element).
   For tokens=1 the i-th item's offset still landed inside the right
   q_dim row by luck of stride, but split-fused with `q_token_offset
   > 0` (the chunked-prefill case) skipped every other token.

   Replaced with `size_of::<half::f16>()` and added a comment so this
   doesn't regress.

+++ crates/ferrum-models/src/models/llama_family.rs
@@ -2503,8 +2503,15 @@ impl<B: Backend> LlamaFamilyModel<B> {
             // Take block_table buffer ptrs ahead of the dispatch loop —
             // we need both per-cache block_table (to write into) and
             // self.scratch.paged_batch_q (to write Q stacks into).
-            let q_head_major_size_bytes = (q_dim * std::mem::size_of::<f32>()) as u64;
-            let qkv_stride_bytes = (qkv_stride * std::mem::size_of::<f32>()) as u64;
+            // f16 buffers (CudaSlice<f16> / Metal half) — 2 bytes per
+            // element. Earlier `size_of::<f32>()` was 2× too large; for
+            // tokens=1 batched decode the per-item Q stack still landed
+            // at the right offset because the caller's stride is q_dim
+            // and i==0..m gives base+i*q_dim, but split-fused with
+            // q_token_offset>0 (chunked prefill) tripped on the bad
+            // offset by skipping every other token row.
+            let q_head_major_size_bytes = (q_dim * std::mem::size_of::<half::f16>()) as u64;
+            let qkv_stride_bytes = (qkv_stride * std::mem::size_of::<half::f16>()) as u64;
             let _t_qkr = if _bp { B::sync(ctx); Some(std::time::Instant::now()) } else { None };
             for (i, (cache_id, _, pos)) in batch.iter().enumerate() {
                 let pos_i = *pos as usize;
@@ -3559,8 +3566,10 @@ impl<B: Backend> LlamaFamilyModel<B> {
         );
         // SAFETY: pools allocated once; not concurrently mutated.
         let (pool_k, pool_v) = unsafe { (&mut *pool_ptr.0, &mut *pool_ptr.1) };
-        let q_head_major_size_bytes = (q_dim * std::mem::size_of::<f32>()) as u64;
-        let qkv_stride_bytes = (qkv_stride * std::mem::size_of::<f32>()) as u64;
+        // f16 buffers — see the comment in `forward_layer_batched_decode_post_attn`
+        // about the prior 2× overshoot from `size_of::<f32>()`.
+        let q_head_major_size_bytes = (q_dim * std::mem::size_of::<half::f16>()) as u64;
+        let qkv_stride_bytes = (qkv_stride * std::mem::size_of::<half::f16>()) as u64;
 
         for (i, (cid, _q_tokens, pos_offset, _)) in items.iter().enumerate() {
             let q_token_offset = cu_seqlens_q[i] as u64;
